+++ src/test/java/com/braintreegateway/integrationtest/PaymentMethodIT.java
@@ -280,6 +280,7 @@ public class PaymentMethodIT extends IntegrationTest {
         assertEquals("checking", usBankAccount.getAccountType());
         assertEquals("PayPal Checking - 1234", usBankAccount.getAccountDescription());
         assertEquals("Dan Schulman", usBankAccount.getAccountHolderName());
+        assertEquals("UNKNOWN", usBankAccount.getBankName());
     }
 
     @Test
@@ -1330,8 +1331,40 @@ public class PaymentMethodIT extends IntegrationTest {
         BraintreeGateway accessTokenGateway = new BraintreeGateway(accessTokenResult.getTarget().getAccessToken());
         Result<PaymentMethodNonce> grantResult = accessTokenGateway.paymentMethod().grant(paymentMethodToken);
         assertTrue(grantResult.isSuccess());
-        NodeWrapper revokeResult = accessTokenGateway.paymentMethod().revoke(paymentMethodToken);
+
+        Result<? extends PaymentMethod> revokeResult = accessTokenGateway.paymentMethod().revoke(paymentMethodToken);
         assertTrue(revokeResult.isSuccess());
     }
 
+    @Test
+    public void grantWithOptionsAndRevoke() {
+        BraintreeGateway partnerMerchantGateway = new BraintreeGateway(Environment.DEVELOPMENT, "integration_merchant_public_id", "oauth_app_partner_user_public_key", "oauth_app_partner_user_private_key");
+        Result<Customer> customerResult = partnerMerchantGateway.customer().create(new CustomerRequest());
+        Customer customer = customerResult.getTarget();
+
+        PaymentMethodRequest request = new PaymentMethodRequest().
+            paymentMethodNonce(Nonce.Transactable).
+            customerId(customer.getId());
+        Result<? extends PaymentMethod> result = partnerMerchantGateway.paymentMethod().create(request);
+        String paymentMethodToken = result.getTarget().getToken();
+        BraintreeGateway oauthGateway = new BraintreeGateway("client_id$development$integration_client_id", "client_secret$development$integration_client_secret");
+        String code = TestHelper.createOAuthGrant(oauthGateway, "integration_merchant_id", "grant_payment_method");
+
+        OAuthCredentialsRequest oauthRequest = new OAuthCredentialsRequest().
+            code(code).
+            scope("grant_payment_method");
+
+        Result<OAuthCredentials> accessTokenResult = oauthGateway.oauth().createTokenFromCode(oauthRequest);
+
+        BraintreeGateway accessTokenGateway = new BraintreeGateway(accessTokenResult.getTarget().getAccessToken());
+
+        PaymentMethodGrantRequest grantRequest = new PaymentMethodGrantRequest().
+            allowVaulting(false).
+            includeBillingPostalCode(true);
+        Result<PaymentMethodNonce> grantResult = accessTokenGateway.paymentMethod().grant(paymentMethodToken, grantRequest);
+        assertTrue(grantResult.isSuccess());
+
+        Result<? extends PaymentMethod> revokeResult = accessTokenGateway.paymentMethod().revoke(paymentMethodToken);
+        assertTrue(revokeResult.isSuccess());
+    }
 }