+++ src/test/java/com/braintreegateway/testhelpers/TestHelper.java
@@ -475,6 +475,7 @@ public abstract class TestHelper {
         try {
             JSONObject json = new JSONObject(clientToken);
             URL url = new URL(json.getJSONObject("braintree_api").getString("url") + "/tokens");
+            String token = json.getJSONObject("braintree_api").getString("access_token");
             SSLContext sc = SSLContext.getInstance("TLSv1.1");
             sc.init(null, null, null);
             HttpsURLConnection connection = (HttpsURLConnection) url.openConnection();
@@ -482,7 +483,7 @@ public abstract class TestHelper {
             connection.setRequestMethod("POST");
             connection.addRequestProperty("Content-Type", "application/json");
             connection.addRequestProperty("Braintree-Version", "2015-11-01");
-            connection.addRequestProperty("Authorization", "Bearer integratexxxxxx_xxxxxx_xxxxxx_xxxxxx_xx1");
+            connection.addRequestProperty("Authorization", "Bearer " + token);
             connection.setDoOutput(true);
             connection.getOutputStream().write(payload.getBytes("UTF-8"));
             connection.getOutputStream().close();
@@ -499,12 +500,13 @@ public abstract class TestHelper {
     }
 
     public static String generateInvalidUsBankAccountNonce() {
-        String[] valid_characters = "bcdfghjkmnpqrstvwxyz23456789".split("");
+        String valid_characters = "bcdfghjkmnpqrstvwxyz23456789";
         String token = "tokenusbankacct";
         for(int i=0; i < 4; i++) {
             token += '_';
             for(int j=0; j<6; j++) {
-               token += valid_characters[new Random().nextInt(valid_characters.length)];
+                Integer pick = new Random().nextInt(valid_characters.length());
+                token += valid_characters.charAt(pick);
             }
         }
         return token + "_xxx";