+++ src/test/java/com/braintreegateway/unittest/ConfigurationTest.java
@@ -1,6 +1,7 @@
 package com.braintreegateway.unittest;
 
 import com.braintreegateway.*;
+import com.braintreegateway.exceptions.ConfigurationException;
 import com.braintreegateway.testhelpers.TestHelper;
 
 import org.junit.Before;
@@ -18,6 +19,36 @@ public class ConfigurationTest {
     public void testStringEnvironmentConstructor() {
        Configuration configuration = new Configuration("development", "merchant_id", "public_key", "private_key");
        assertEquals(Environment.DEVELOPMENT, configuration.getEnvironment());
+    }
+
+    @Test(expected=ConfigurationException.class)
+    public void configurationThrowsErrorOnNullMerchantId() {
+        Configuration configuration = new Configuration(Environment.DEVELOPMENT, null, "integration_public_key", "integration_private_key");
+    }
+
+    @Test(expected=ConfigurationException.class)
+    public void configurationThrowsErrorOnEmptyMerchantId() {
+        Configuration configuration = new Configuration(Environment.DEVELOPMENT, "", "integration_public_key", "integration_private_key");
+    }
+
+    @Test(expected=ConfigurationException.class)
+    public void configurationThrowsErrorOnNullPublicKey() {
+        Configuration configuration = new Configuration(Environment.DEVELOPMENT, "merchant_id", null, "integration_private_key");
+    }
+
+    @Test(expected=ConfigurationException.class)
+    public void configurationThrowsErrorOnEmptyPublicKey() {
+        Configuration configuration = new Configuration(Environment.DEVELOPMENT, "merchant_id", "", "integration_private_key");
+    }
+
+    @Test(expected=ConfigurationException.class)
+    public void configurationThrowsErrorOnNullPrivateKey() {
+        Configuration configuration = new Configuration(Environment.DEVELOPMENT, "merchant_id", "integration_public_key", null);
+    }
+
+    @Test(expected=ConfigurationException.class)
+    public void configurationThrowsErrorOnEmptyPrivateKey() {
+        Configuration configuration = new Configuration(Environment.DEVELOPMENT, "merchant_id", "integration_public_key", "");
     }
 
     @Test